Organize modules into mailmanager package
